requirements.txt
arrow==1.3.0
geopandas==0.13.2
matplotlib==3.8.0
networkx==3.2
numpy==1.26.1
pandas==2.1.4
plotly==5.17.0
pyarrow==13.0.0
pyarrow-hotfix==0.6
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
xgboost
hyperopt
shap
neptune
statsmodels
lightgbm
python-dotenv
boto3